// src/interfaces/address.ts
export interface Address {
    firstName: string;
    lastName: string;
    address: string;
    address2: string;
    postalCode: string;
    city: string;
    country: string;
    phone: string;
  }

  export interface OrderAddress {
    firstName: string;
    lastName: string;
    address: string;
    address2: string;
    postalCode: string;
    city: string;
    countryId: string;
    phone: string;
  }